feat: Refactor Conversation Simulator app into modular structure with dedicated helper functions
